Add files via upload
This is the hdmi config file for using the Raspberry pi with the Adafruit 7 inch hdmi screen with touch backpack. For the touchscreen to work the usb cable on the hdmi screen must be connected to one of the usb ports on the Pi itself. 
Adafruit overview: https://learn.adafruit.com/adafruit-5-800x480-tft-hdmi-monitor-touchscreen-backpack
USB plugged in source: https://forums.adafruit.com/viewtopic.php?f=47&t=88040